+++ examples/js/manipulation/export-csv-table.js
@@ -1,4 +1,5 @@
 /* eslint max-len: 0 */
+/* eslint no-unused-vars: 0*/
 import React from 'react';
 import { BootstrapTable, TableHeaderColumn } from 'react-bootstrap-table';
 
@@ -21,8 +22,8 @@ addProducts(5);
 
 export default class ExportCSVTable extends React.Component {
 
-  csvFormatter(cell) {
-    return cell + ' USD';
+  csvFormatter(cell, row) {
+    return `${row.id}: ${cell} USD`;
   }
 
   render() {